+++ packages/superset-ui-plugin-chart-word-cloud/src/chart/Encoder.ts
