+++ packages/superset-ui-plugin-chart-word-cloud/src/legacy/index.ts
@@ -17,7 +17,7 @@ export default class WordCloudChartPlugin extends ChartPlugin<LegacyWordCloudFor
   constructor() {
     super({
       buildQuery,
-      loadChart: () => import('../WordCloud'),
+      loadChart: () => import('../chart/WordCloud'),
       metadata,
       transformProps,
     });
