+++ packages/superset-ui-plugins-demo/storybook/stories/plugin-chart-word-cloud/index.js
@@ -1,4 +1,4 @@
-import WordCloudChartPlugin from '../../../../superset-ui-plugin-chart-word-cloud';
+import WordCloudChartPlugin from '../../../../superset-ui-plugin-chart-word-cloud/src';
 import LegacyWordCloudChartPlugin from '../../../../superset-ui-plugin-chart-word-cloud/esm/legacy';
 import Stories from './Stories';
